update the mapping resource and the workflow
The Simple Standard for Sharing Ontology
Mappings (SSSOM) is used to form metadata of a mapping. The SKOS:closeMatch is used as the predicate id of all the mappings. The VoID and extended DCAT2 (with mapping objectproperty and Mapping class) is used to form metadata schema of the resource of mappings.